requirement.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.0.0
plotly>=5.15.0
PyYAML>=6.0.0
openpyxl>=3.1.0
kaleido>=0.2.1
diskcache>=5.6.0
pyarrow>=12.0.0
scipy>=1.10.0
python-dotenv>=1.0.0
reportlab>=4.0.0